backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
chromadb==0.4.24
httpx==0.26.0
python-dotenv==1.0.0
pandas>=2.2.0
sentence-transformers>=2.2.2
langchain-community==0.0.19
